+++ mobile/android/android-components/buildSrc/src/main/java/Gecko.kt
@@ -6,17 +6,17 @@ internal object GeckoVersions {
     /**
      * GeckoView Nightly Version.
      */
-    const val nightly_version = "76.0.20200406092400"
+    const val nightly_version = "77.0.20200407093520"
 
     /**
      * GeckoView Beta Version.
      */
-    const val beta_version = "75.0.20200403064004"
+    const val beta_version = "76.0.20200406152003"
 
     /**
      * GeckoView Release Version.
      */
-    const val release_version = "74.0.20200309095159"
+    const val release_version = "75.0.20200403170909"
 }
 
 @Suppress("MaxLineLength")