+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -16,9 +16,9 @@ import mozilla.components.browser.engine.gecko.webpush.GeckoWebPushDelegate
 import mozilla.components.browser.engine.gecko.webpush.GeckoWebPushHandler
 import mozilla.components.concept.engine.Engine
 import mozilla.components.concept.engine.EngineSession
+import mozilla.components.concept.engine.EngineSession.SafeBrowsingPolicy
 import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy
 import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy.TrackingCategory
-import mozilla.components.concept.engine.EngineSession.SafeBrowsingPolicy
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.engine.Settings
@@ -30,6 +30,7 @@ import mozilla.components.concept.engine.utils.EngineVersion
 import mozilla.components.concept.engine.webextension.Action
 import mozilla.components.concept.engine.webextension.ActionHandler
 import mozilla.components.concept.engine.webextension.EnableSource
+import mozilla.components.concept.engine.webextension.TabHandler
 import mozilla.components.concept.engine.webextension.WebExtension
 import mozilla.components.concept.engine.webextension.WebExtensionDelegate
 import mozilla.components.concept.engine.webextension.WebExtensionRuntime
@@ -44,11 +45,9 @@ import org.mozilla.geckoview.ContentBlockingController
 import org.mozilla.geckoview.ContentBlockingController.Event
 import org.mozilla.geckoview.GeckoResult
 import org.mozilla.geckoview.GeckoRuntime
-import org.mozilla.geckoview.GeckoRuntimeSettings
 import org.mozilla.geckoview.GeckoSession
 import org.mozilla.geckoview.GeckoWebExecutor
 import org.mozilla.geckoview.WebExtensionController
-import java.lang.IllegalStateException
 
 /**
  * Gecko-based implementation of Engine interface.
@@ -78,6 +77,11 @@ class GeckoEngine(
 
         override fun onToggleActionPopup(extension: WebExtension, action: Action): EngineSession? {
             return webExtensionDelegate?.onToggleActionPopup(extension, GeckoEngineSession(runtime), action)
+        }
+    }
+    private val webExtensionTabHandler = object : TabHandler {
+        override fun onNewTab(webExtension: WebExtension, engineSession: EngineSession, active: Boolean, url: String) {
+            webExtensionDelegate?.onNewTab(webExtension, engineSession, active, url)
         }
     }
 
@@ -193,7 +197,7 @@ class GeckoEngine(
         onSuccess: ((WebExtension) -> Unit),
         onError: ((String, Throwable) -> Unit)
     ) {
-        val ext = GeckoWebExtension(id, url, runtime.webExtensionController, allowContentMessaging, supportActions)
+        val ext = GeckoWebExtension(id, url, runtime, allowContentMessaging, supportActions)
         installWebExtension(ext, onSuccess, onError)
     }
 
@@ -213,6 +217,7 @@ class GeckoEngine(
                 // https://bugzilla.mozilla.org/show_bug.cgi?id=1599897
                 // https://bugzilla.mozilla.org/show_bug.cgi?id=1582185
                 ext.registerActionHandler(webExtensionActionHandler)
+                ext.registerTabHandler(webExtensionTabHandler)
             }
 
             // For now we have to use registerWebExtension for builtin extensions until we get the
@@ -227,9 +232,10 @@ class GeckoEngine(
             })
         } else {
             runtime.webExtensionController.install(ext.url).then({
-                val installedExtension = GeckoWebExtension(it!!, runtime.webExtensionController)
+                val installedExtension = GeckoWebExtension(it!!, runtime)
                 webExtensionDelegate?.onInstalled(installedExtension)
                 installedExtension.registerActionHandler(webExtensionActionHandler)
+                installedExtension.registerTabHandler(webExtensionTabHandler)
                 onSuccess(installedExtension)
                 GeckoResult<Void>()
             }, { throwable ->
@@ -267,8 +273,9 @@ class GeckoEngine(
     ) {
         runtime.webExtensionController.update((extension as GeckoWebExtension).nativeExtension).then({ geckoExtension ->
             val updatedExtension = if (geckoExtension != null) {
-                GeckoWebExtension(geckoExtension, runtime.webExtensionController).also {
+                GeckoWebExtension(geckoExtension, runtime).also {
                     it.registerActionHandler(webExtensionActionHandler)
+                    it.registerTabHandler(webExtensionTabHandler)
                 }
             } else {
                 null
@@ -284,6 +291,7 @@ class GeckoEngine(
     /**
      * See [Engine.registerWebExtensionDelegate].
      */
+    @Suppress("Deprecation")
     override fun registerWebExtensionDelegate(
         webExtensionDelegate: WebExtensionDelegate
     ) {
@@ -291,7 +299,7 @@ class GeckoEngine(
 
         val promptDelegate = object : WebExtensionController.PromptDelegate {
             override fun onInstallPrompt(ext: org.mozilla.geckoview.WebExtension): GeckoResult<AllowOrDeny>? {
-                val extension = GeckoWebExtension(ext, runtime.webExtensionController)
+                val extension = GeckoWebExtension(ext, runtime)
                 return if (webExtensionDelegate.onInstallPermissionRequest(extension)) {
                     GeckoResult.ALLOW
                 } else {
@@ -308,8 +316,8 @@ class GeckoEngine(
                 // NB: We don't have a user flow for handling updated origins so we ignore them for now.
                 val result = GeckoResult<AllowOrDeny>()
                 webExtensionDelegate.onUpdatePermissionRequest(
-                    GeckoWebExtension(current, runtime.webExtensionController),
-                    GeckoWebExtension(updated, runtime.webExtensionController),
+                    GeckoWebExtension(current, runtime),
+                    GeckoWebExtension(updated, runtime),
                     newPermissions.toList()
                 ) {
                     allow -> if (allow) result.complete(AllowOrDeny.ALLOW) else result.complete(AllowOrDeny.DENY)
@@ -318,7 +326,14 @@ class GeckoEngine(
             }
         }
 
+        val debuggerDelegate = object : WebExtensionController.DebuggerDelegate {
+            override fun onExtensionListUpdated() {
+                webExtensionDelegate.onExtensionListUpdated()
+            }
+        }
+
         runtime.webExtensionController.promptDelegate = promptDelegate
+        runtime.webExtensionController.setDebuggerDelegate(debuggerDelegate)
     }
 
     /**
@@ -327,10 +342,13 @@ class GeckoEngine(
     override fun listInstalledWebExtensions(onSuccess: (List<WebExtension>) -> Unit, onError: (Throwable) -> Unit) {
         runtime.webExtensionController.list().then({
             val extensions = it?.map {
-                extension -> GeckoWebExtension(extension, runtime.webExtensionController)
+                extension -> GeckoWebExtension(extension, runtime)
             } ?: emptyList()
 
-            extensions.forEach { extension -> extension.registerActionHandler(webExtensionActionHandler) }
+            extensions.forEach { extension ->
+                extension.registerActionHandler(webExtensionActionHandler)
+                extension.registerTabHandler(webExtensionTabHandler)
+            }
             onSuccess(extensions)
             GeckoResult<Void>()
         }, { throwable ->
@@ -349,7 +367,7 @@ class GeckoEngine(
         onError: (Throwable) -> Unit
     ) {
         runtime.webExtensionController.enable((extension as GeckoWebExtension).nativeExtension, source.id).then({
-            val enabledExtension = GeckoWebExtension(it!!, runtime.webExtensionController)
+            val enabledExtension = GeckoWebExtension(it!!, runtime)
             webExtensionDelegate?.onEnabled(enabledExtension)
             onSuccess(enabledExtension)
             GeckoResult<Void>()
@@ -369,7 +387,7 @@ class GeckoEngine(
         onError: (Throwable) -> Unit
     ) {
         runtime.webExtensionController.disable((extension as GeckoWebExtension).nativeExtension, source.id).then({
-            val disabledExtension = GeckoWebExtension(it!!, runtime.webExtensionController)
+            val disabledExtension = GeckoWebExtension(it!!, runtime)
             webExtensionDelegate?.onDisabled(disabledExtension)
             onSuccess(disabledExtension)
             GeckoResult<Void>()
@@ -435,13 +453,6 @@ class GeckoEngine(
      * See [Engine.settings]
      */
     override val settings: Settings = object : Settings() {
-
-        init {
-            // `autoplayDefault` has been removed from GV nightly, but for now it must be set to
-            // `allowed` in order for site specific autoplay permissions to work.
-            runtime.settings.autoplayDefault = GeckoRuntimeSettings.AUTOPLAY_DEFAULT_ALLOWED
-        }
-
         override var javascriptEnabled: Boolean
             get() = runtime.settings.javaScriptEnabled
             set(value) { runtime.settings.javaScriptEnabled = value }
@@ -574,7 +585,6 @@ class GeckoEngine(
             this.testingModeEnabled = it.testingModeEnabled
             this.userAgentString = it.userAgentString
             this.preferredColorScheme = it.preferredColorScheme
-            this.suspendMediaWhenInactive = it.suspendMediaWhenInactive
             this.fontInflationEnabled = it.fontInflationEnabled
             this.fontSizeFactor = it.fontSizeFactor
             this.forceUserScalableContent = it.forceUserScalableContent