+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -145,6 +145,7 @@ class GeckoEngineSession(
             notifyObservers { onNavigateBack() }
         }
     }
+
     /**
      * See [EngineSession.goForward]
      */
@@ -465,7 +466,7 @@ class GeckoEngineSession(
 
             notifyObservers {
                 // TODO provide full certificate info: https://github.com/mozilla-mobile/android-components/issues/5557
-                onSecurityChange(securityInfo.isSecure, securityInfo.host, securityInfo.certificate?.issuerDN?.name)
+                onSecurityChange(securityInfo.isSecure, securityInfo.host, securityInfo.getIssuerName())
             }
         }
 
@@ -752,6 +753,10 @@ class GeckoEngineSession(
         }
 
         return cookiesPolicies
+    }
+
+    internal fun GeckoSession.ProgressDelegate.SecurityInformation.getIssuerName(): String? {
+        return certificate?.issuerDN?.name?.substringAfterLast("O=")?.substringBeforeLast(",C=")
     }
 
     private operator fun Int.contains(mask: Int): Boolean {